Review fix

+++ server/Models/PcBuild.cs
@@ -5,7 +5,7 @@ public class PcBuild : RepoItem<int>{
     public string CreatorId {get; set;}
     public int PowerScore {get; set;}
     public decimal Price {get; set;}
-    public int Rating {get; set;}
+    public decimal Rating {get; set;}
     public Boolean IsPrivate {get; set;}
     public string PcCase {get; set;}
     public string CasePicture {get; set;}
